Added an object to simplify working with location object
"new URLSearchParams" is getting "location.search" this simplifies working with location object. Once new object is created then ".get()" method is run. The method takes 'sort' of the URL which can be compared with 'asc' and the result of comparison returns either "true" or "false". This value is fianlly used to dynamically change URL by swapping "desc" or "asc".