Added funtionality to fetch data from Google Firebase
Custom "useHttp" hook was used and "getAllquotes" function is passed as an argument. "useHttp" then returns various constants such as "sendRequest", "status" and so on. In Addition, "useEffect" is used to send request to Google Firebase RealTime database and when the "status" is received it is then checked and according to various states different action are taken.