Extracting a singlequote from Google Firebas
For fetching a single quote from Google Firebase RealTime DB custom "useHttp" is used. "useHttp" takes "getSingleQuote" function as an argument and it then returns "sendRequest", "status", "error" and "loadQuote". "sendRequest" is used for sending a request and it takes "quoteId" as argument. The "sendRequest" is placed inside "useEffect" which ensure the request is sent after the components loads. There also three "if" checks that shows different components. Finally quote is loaded if there are no issues.